Fix MaxFEM compilation and style errors

+++ src/core/model/sequence/timestamp.h
@@ -2,6 +2,6 @@
 
 namespace model {
 
-using Timestamp = unsigned int;
+using Timestamp = uint64_t;
 
 }  // namespace model
